# src/ad-serving/dto/track-event.dto.ts
import { IsUUID, IsNotEmpty } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class TrackEventDto {
  @ApiProperty({
    format: 'uuid',
    example: '123e4567-e89b-12d3-a456-426614174000',
    description: 'ID of the ad to track',
  })
  @IsUUID()
  @IsNotEmpty()
  adId: string;
}